3.2

+++ index.js
@@ -13,12 +13,11 @@ dotenv.config()
 
 // kanāli kuros ziņas izdzēšas pēc noteikta laika
 const channels = [
-  '797584379685240885',
-  '797587282672484392',
-  '890662723648647249'
+  '898550678417457163', // klacukanals
+  '797587282672484392', // sudpublicesana
+  '890662723648647249', // politika
+  '875083366611955715'  // test kanals
 ]
-
-const timeout = 300000
 
 // definē DiscordJS klientu
 const client = new Client({
@@ -51,13 +50,17 @@ client.on('ready', async () => {
   await cacheKaktus()
   await kaktsLoop()
 
+  const timeout = 20000
+
   client.on("messageCreate", async message => {
     // pārbauda vai ziņa nav no bota
     if (message.author.id === process.env.ULMANISID) {
       channels.map(channel => {
-        if (channel === message.channelId) setTimeout(() => {
-          message.delete()
-        }, timeout)
+        if (channel === message.channelId && message.embeds.length) {
+          setTimeout(() => {
+            if (!message.deleted) message.delete()
+          }, timeout)
+        }
       })
     } else {
       await antispam(message)
@@ -79,9 +82,7 @@ export const kaktsRole = async (guildId, userId, isAdd = 1) => {
   const guild = await client.guilds.cache.get(guildId)
   const kakts = await guild.roles.cache.get(kaktsRoleId)
   const member = await guild.members.cache.get(userId)
-  if (isAdd) {
-    await member.roles.add(kakts)
-  } else {
-    await member.roles.remove(kakts)
-  }
+
+  if (isAdd) await member.roles.add(kakts)
+  else await member.roles.remove(kakts)
 }
